fix: moove common things into common file
