+++ src/features/defi/providers/foxy/components/FoxyManager/Deposit/DepositContext.ts
@@ -1,6 +1,6 @@
 import { createContext } from 'react'
 
-import { FoxyDepositActions, FoxyDepositState } from './DepositReducer'
+import { FoxyDepositActions, FoxyDepositState } from './DepositCommon'
 
 export interface IDepositContext {
   state: FoxyDepositState | null
