+++ src/features/defi/providers/foxy/components/FoxyManager/Deposit/components/Approve.tsx
@@ -18,10 +18,10 @@ import { selectAssetByCAIP19, selectMarketDataById } from 'state/slices/selector
 import { useAppDispatch, useAppSelector } from 'state/store'
 
 import { DepositPath } from '../DepositCommon'
+import { FoxyDepositActionType } from '../DepositCommon'
 import { DepositContext } from '../DepositContext'
-import { FoxyDepositActionType } from '../DepositReducer'
 
-export type FoxyApproveProps = {
+type FoxyApproveProps = {
   api: FoxyApi
   getDepositGasEstimate: (deposit: DepositValues) => Promise<string | undefined>
 }
