+++ src/features/defi/providers/yearn/components/YearnManager/Deposit/DepositReducer.ts
@@ -1,26 +1,6 @@
-import { SupportedYearnVault, YearnVault } from '@shapeshiftoss/investor-yearn'
 import { ChainTypes } from '@shapeshiftoss/types'
-import { DepositValues } from 'features/defi/components/Deposit/Deposit'
 
-type EstimatedGas = {
-  estimatedGasCrypto?: string
-}
-
-type YearnDepositValues = DepositValues &
-  EstimatedGas & {
-    txStatus: string
-    usedGasFee: string
-  }
-
-export type YearnDepositState = {
-  vault: SupportedYearnVault
-  userAddress: string | null
-  approve: EstimatedGas
-  deposit: YearnDepositValues
-  loading: boolean
-  pricePerShare: string
-  txid: string | null
-}
+import { YearnDepositActions, YearnDepositActionType, YearnDepositState } from './DepositCommon'
 
 export const initialState: YearnDepositState = {
   txid: null,
@@ -71,61 +51,6 @@ export const initialState: YearnDepositState = {
     usedGasFee: '',
   },
 }
-
-export enum YearnDepositActionType {
-  SET_VAULT = 'SET_VAULT',
-  SET_APPROVE = 'SET_APPROVE',
-  SET_USER_ADDRESS = 'SET_USER_ADDRESS',
-  SET_DEPOSIT = 'SET_DEPOSIT',
-  SET_LOADING = 'SET_LOADING',
-  SET_PRICE_PER_SHARE = 'SET_PRICE_PER_SHARE',
-  SET_TXID = 'SET_TXID',
-  SET_TX_STATUS = 'SET_TX_STATUS',
-}
-
-type SetVaultAction = {
-  type: YearnDepositActionType.SET_VAULT
-  payload: YearnVault | null
-}
-
-type SetApprove = {
-  type: YearnDepositActionType.SET_APPROVE
-  payload: EstimatedGas
-}
-
-type SetDeposit = {
-  type: YearnDepositActionType.SET_DEPOSIT
-  payload: Partial<YearnDepositValues>
-}
-
-type SetUserAddress = {
-  type: YearnDepositActionType.SET_USER_ADDRESS
-  payload: string
-}
-
-type SetLoading = {
-  type: YearnDepositActionType.SET_LOADING
-  payload: boolean
-}
-
-type SetPricePerShare = {
-  type: YearnDepositActionType.SET_PRICE_PER_SHARE
-  payload: string
-}
-
-type SetTxid = {
-  type: YearnDepositActionType.SET_TXID
-  payload: string
-}
-
-export type YearnDepositActions =
-  | SetVaultAction
-  | SetApprove
-  | SetDeposit
-  | SetUserAddress
-  | SetLoading
-  | SetPricePerShare
-  | SetTxid
 
 export const reducer = (state: YearnDepositState, action: YearnDepositActions) => {
   switch (action.type) {
