+++ src/features/defi/providers/yearn/components/YearnManager/Deposit/components/Confirm.tsx
@@ -21,11 +21,10 @@ import { marketApi } from 'state/slices/marketDataSlice/marketDataSlice'
 import { selectAssetByCAIP19, selectMarketDataById } from 'state/slices/selectors'
 import { useAppDispatch, useAppSelector } from 'state/store'
 
-import { DepositPath } from '../DepositCommon'
+import { DepositPath, YearnDepositActionType } from '../DepositCommon'
 import { DepositContext } from '../DepositContext'
-import { YearnDepositActionType } from '../DepositReducer'
 
-export type YearnConfirmProps = {
+type YearnConfirmProps = {
   api: YearnVaultApi
 }
