+++ src/features/defi/providers/yearn/components/YearnManager/Withdraw/components/Confirm.tsx
@@ -20,9 +20,8 @@ import { poll } from 'lib/poll/poll'
 import { selectAssetByCAIP19, selectMarketDataById } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
-import { WithdrawPath } from '../WithdrawCommon'
+import { WithdrawPath, YearnWithdrawActionType } from '../WithdrawCommon'
 import { WithdrawContext } from '../WithdrawContext'
-import { YearnWithdrawActionType } from '../WithdrawReducer'
 
 type YearnConfirmProps = {
   api: YearnVaultApi
